Refactor the state machine to reutilize more components
