+++ kolibri/plugins/user_profile/assets/test/views/ChangeFacility/UsernameExists.spec.js
@@ -38,10 +38,15 @@ describe(`ChangeFacility/UsernameExists`, () => {
   it(`Show correct info`, () => {
     const wrapper = makeWrapper({
       targetFacility: { name: 'Test Facility' },
-      username: 'Test User',
+      username: 'test1',
     });
-    expect(wrapper.text()).toContain(
-      'Change Facility An account with the username ‘Test User’ already exists in the ‘Test Facility’ learning facility.'
+    const line1Paragraph = wrapper.find('[data-test="line1"]');
+    expect(line1Paragraph.text()).toEqual(
+      'An account with the username ‘test1’ already exists in the ‘Test Facility’ learning facility.'
+    );
+    const line2Paragraph = wrapper.find('[data-test="line2"]');
+    expect(line2Paragraph.text()).toEqual(
+      'You can merge all of your account and progress data with this account in ‘Test Facility’ learning facility, or you can create a new account. All of your progress data will be moved to this new account.'
     );
   });
   it(`shows the buttons`, () => {
